+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/visual/renderables/NewLineRenderable.java
@@ -81,7 +81,7 @@ public class NewLineRenderable implements GLRenderable {
 
     @Override
     public int getPriority() {
-        return GLRenderable.ANNOTATIONS_PRIORITY;
+        return RenderablePriority.ANNOTATIONS_PRIORITY.getValue();
     }
 
     /**