+++ CoreOpenGLDisplay/src/au/gov/asd/tac/constellation/visual/opengl/renderer/AxesRenderable.java
@@ -78,7 +78,7 @@ public class AxesRenderable implements GLRenderable {
 
     @Override
     public int getPriority() {
-        return GLRenderable.ANNOTATIONS_PRIORITY;
+        return RenderablePriority.ANNOTATIONS_PRIORITY.getValue();
     }
 
     @Override